user and Driver Rate

+++ Models/RateDriver.cs
@@ -10,7 +10,7 @@ namespace UMoveNew.Models
         public int ID { get; set; }
         public int TripID { get; set; }
         public int DriverID { get; set; }
-        public int Rate { get; set; }
+        public decimal Rate { get; set; }
 
         public string Name { get; set; }
         public int UserID { get; set; }
